Dynamic sdk instantiation
