+++ packages/react-sdk/src/hooks/useEthereum.ts
@@ -6,7 +6,15 @@ import { usePhantom } from "../PhantomProvider";
  * @returns Ethereum chain interface with connection enforcement
  */
 export function useEthereum() {
-  const { sdk, isConnected } = usePhantom();
+  const { sdk, isConnected, isClient } = usePhantom();
+
+  if (!isClient || !sdk) {
+    // Return a stub object for SSR or while SDK is initializing
+    return {
+      ethereum: {} as any, // This will be replaced when SDK is ready
+      isAvailable: false,
+    };
+  }
 
   return {
     // Chain instance with connection enforcement for signing methods
