+++ packages/react-sdk/src/hooks/useSolana.ts
@@ -6,7 +6,15 @@ import { usePhantom } from "../PhantomProvider";
  * @returns Solana chain interface with connection enforcement
  */
 export function useSolana() {
-  const { sdk, isConnected } = usePhantom();
+  const { sdk, isConnected, isClient } = usePhantom();
+
+  if (!isClient || !sdk) {
+    // Return a stub object for SSR or while SDK is initializing
+    return {
+      solana: {} as any, // This will be replaced when SDK is ready
+      isAvailable: false,
+    };
+  }
 
   return {
     // Chain instance with connection enforcement for signing methods
